requirements.txt
mock
unittest2
boto
dnspython
fabric
